Made --help more helpful.
Now hiding the commented wpa2 command and help behind #ifdef WPA2_MODE.

+++ src/dl.c
@@ -58,10 +58,10 @@ int dl(const char *path, FILE *fp)
 		errmsg = "`BEGIN_UPLOAD` was denied.";
 		return ERR_VM;
 	}
-	unsigned read_so_far = burep->packetLen + 2 - offsetof(BEGIN_UPLOAD_REPLY, bytes);
+	unsigned int read_so_far = burep->packetLen + 2 - (unsigned int) offsetof(BEGIN_UPLOAD_REPLY, bytes);
 	fwrite(burep->bytes, read_so_far, 1, fp);
 
-	unsigned total = burep->fileSize;
+	unsigned int total = burep->fileSize;
 
 	CONTINUE_UPLOAD cu = CONTINUE_UPLOAD_INIT;
 	cu.fileHandle = burep->fileHandle;
